ui_service/requirements.txt
streamlit
kafka-python
psycopg2-binary
pandas
sqlalchemy
